+++ lib/tests/test_default_revset_graph_iterator.rs
@@ -17,7 +17,8 @@ use jj_lib::commit::Commit;
 use jj_lib::default_index_store::ReadonlyIndexWrapper;
 use jj_lib::default_revset_engine::{evaluate, RevsetImpl};
 use jj_lib::repo::{ReadonlyRepo, Repo as _};
-use jj_lib::revset::{ResolvedExpression, RevsetGraphEdge};
+use jj_lib::revset::ResolvedExpression;
+use jj_lib::revset_graph::RevsetGraphEdge;
 use test_case::test_case;
 use testutils::{CommitGraphBuilder, TestRepo};
